+++ Assets/AXIP/AILIA-MODELS/AudioProcessing/AiliaMicrophone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f23a4d50bac542e3bd8dc77ab7cd3ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
